:sparkles: feat: adds validation to backend and feedbacks on UI

+++ backend/src/controllers/task_controller.ts
@@ -1,7 +1,6 @@
 import AppDataSource from "../config/db_data_source";
 import { FastifyRequest, FastifyReply } from "fastify";
 import { Task } from "../entities/task";
-import UpdateTaskBody from "../models/update_task_body";
 import { TaskStatus } from "../entities/task_status";
 
 export class TaskController {
@@ -22,17 +21,25 @@ export class TaskController {
     static async store(req: FastifyRequest<{ Body: { userId: string, title: string, description: string, statusId?: string } }>, reply: FastifyReply) {
         const { userId, title, description, statusId } = req.body;
 
+        if (!userId || !title || !description || !statusId) {
+            return reply.status(400).send({
+                error: "UserId, title, description and statusId are required."
+            });
+        }
+
+        if (userId.trim().length === 0 || title.trim().length === 0 || description.trim().length === 0 || statusId.trim().length === 0) {
+            return reply.status(400).send({
+                error: "UserId, title, description and statusId are required."
+            });
+        }
+
         const taskRepository = AppDataSource.getRepository(Task);
         const statusRepository = AppDataSource.getRepository(TaskStatus);
 
         let status: TaskStatus | null = null;
 
-        if (!statusId) {
-            return reply.code(400).send({ error: "StatusId é obrigatório" });
-        }
-
         status = await statusRepository.findOne({ where: { id: statusId } });
-        if (!status) return reply.code(400).send({ error: "Status inválido" });
+        if (!status) return reply.code(404).send({ error: "Could not find status" });
 
         const task = taskRepository.create({
             title,
@@ -41,23 +48,35 @@ export class TaskController {
             status,
         });
         await taskRepository.save(task);
-        return reply.status(200).send(task);
+        return reply.status(201).send(task);
     }
 
     static async update(req: FastifyRequest<{ Params: { id: string }; Body: { title?: string, description?: string, statusId?: string } }>, reply: FastifyReply) {
         const { id } = req.params;
         const { title, description, statusId } = req.body;
+
+        if (!id || !title || !description || !statusId) {
+            return reply.code(400).send({ error: "Id, title, description and statusId are required" });
+        }
+
+        if (id.trim().length === 0 || title.trim().length === 0 || description.trim().length === 0 || statusId.trim().length === 0) {
+            return reply.code(400).send({ error: "Id, title, description and statusId are required" });
+        }
+
         const taskRepository = AppDataSource.getRepository(Task);
         const statusRepository = AppDataSource.getRepository(TaskStatus);
+
         const task = await taskRepository.findOne({ where: { id }, relations: ["status"] });
+
         if (!task) {
-            return reply.code(404).send({ error: "Task não encontrada" });
+            return reply.code(404).send({ error: "Task not found" });
         }
+
         if (title !== undefined) task.title = title;
         if (description !== undefined) task.description = description;
         if (statusId !== undefined) {
             const status = await statusRepository.findOne({ where: { id: statusId } });
-            if (!status) return reply.code(400).send({ error: "Status inválido" });
+            if (!status) return reply.code(404).send({ error: "statusId is invalid" });
             task.status = status;
         }
         await taskRepository.save(task);
@@ -66,10 +85,17 @@ export class TaskController {
 
     static async delete(req: FastifyRequest<{ Params: { id: string } }>, reply: FastifyReply) {
         const { id } = req.params;
+        if (!id) {
+            return reply.code(400).send({ error: "Id is required" });
+        }
+        if (id.trim().length === 0) {
+            return reply.code(400).send({ error: "Id is required" });
+        }
+
         const taskRepository = AppDataSource.getRepository(Task);
         const task = await taskRepository.findOne({ where: { id } });
         if (!task) {
-            return reply.code(404).send({ error: "Task não encontrada" });
+            return reply.code(404).send({ error: "Task not found" });
         }
         await taskRepository.remove(task);
         return reply.code(204).send();
